refactor: Move common CLI and Gradle logic into lib and simplify Gradle validate task
