+++ modules/lib/src/main/kotlin/dev/vulnlog/lib/shell/ValidateFile.kt
@@ -1,10 +1,7 @@
 // Copyright 2024 the Vulnlog contributors
 // SPDX-License-Identifier: Apache-2.0
-package dev.vulnlog.cli.shell.shared
+package dev.vulnlog.lib.shell
 
-import com.github.ajalt.clikt.core.CliktCommand
-import com.github.ajalt.clikt.core.ProgramResult
-import dev.vulnlog.cli.shell.ExitCode
 import dev.vulnlog.lib.core.renderValidation
 import dev.vulnlog.lib.core.validate
 import dev.vulnlog.lib.model.VulnlogFileContext
@@ -13,20 +10,7 @@ import dev.vulnlog.lib.result.ValidationResult
 import dev.vulnlog.lib.result.ValidationResults
 import java.io.File
 
-fun CliktCommand.validateParsedInputOrFailWithFailureOutput(
-    fileToResult: Map<File, ParseResult.Ok>,
-): ValidationResults {
-    val validationFindings = validateFiles(fileToResult)
-    if (validationFindings.renderedFindings.isNotBlank()) {
-        echo(validationFindings.renderedFindings, err = true)
-    }
-    if (validationFindings.hasErrors) {
-        throw ProgramResult(ExitCode.VALIDATION_ERROR.ordinal)
-    }
-    return validationFindings
-}
-
-private fun validateFiles(fileToResult: Map<File, ParseResult.Ok>): ValidationResults {
+fun validateFiles(fileToResult: Map<File, ParseResult.Ok>): ValidationResults {
     val contextToResults = validateEachFile(fileToResult)
     val renderedFindings =
         contextToResults
